Separated service API & implementation
LoginService class, which consisted of the declaration and definition of
service method is now separated into an interface for method declaration
and an implementation for method definition.
