+++ service/src/main/java/com/oasis/service/api/LoginServiceApi.java
@@ -1,0 +1,7 @@
+package com.oasis.service.api;
+
+import com.oasis.model.entity.EmployeeModel;
+
+public interface LoginServiceApi {
+    public EmployeeModel checkLoginCredentials(String username, String password);
+}
