+++ service/src/main/java/com/oasis/service/implementation/LoginServiceImpl.java
@@ -1,16 +1,18 @@
-package com.oasis.service;
+package com.oasis.service.implementation;
 
 import com.oasis.model.entity.EmployeeModel;
 import com.oasis.repository.LoginRepository;
+import com.oasis.service.api.LoginServiceApi;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 @Service
-public class LoginService {
+public class LoginServiceImpl implements LoginServiceApi {
     @Autowired
     private LoginRepository loginRepository;
 
+    @Override
     public EmployeeModel checkLoginCredentials(String username, String password) {
         EmployeeModel result = loginRepository.findByUsername(username);
